backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
psycopg2
python-dotenv
openai
requests
google-generativeai
flask-restx
gunicorn
agora-token-builder
redis
